reorganise IIIF packages
