+++ IIIF/Presentation/Annotation/IAnnotation.cs
@@ -1,4 +1,4 @@
-﻿namespace IIIF.Presentation
+﻿namespace IIIF.Presentation.Annotation
 {
     public interface IAnnotation
     {
