+++ IIIF/Presentation/SpecificResource.cs
@@ -1,9 +1,11 @@
-﻿namespace IIIF.Presentation
+﻿using IIIF.Presentation.Selectors;
+
+namespace IIIF.Presentation
 {
     public class SpecificResource : ResourceBase, IStructuralLocation
     {
         public override string Type => nameof(SpecificResource);
         public string Source { get; set; }
-        public ISelector Selector { get; set; } // TODO... Point and Fragment Selector impls?
+        public ISelector Selector { get; set; } 
     }
 }
