+++ IIIF/Presentation/StructureBase.cs
@@ -1,6 +1,6 @@
-﻿using System;
+﻿using IIIF.Presentation.Annotation;
+using System;
 using System.Collections.Generic;
-using System.Text;
 
 namespace IIIF.Presentation
 {
